move CategoryUiState model to shared package
