+++ trends_presentation/src/commonMain/kotlin/net/thechance/mena/trends/presentation/shared/component/CategoryItem.kt
@@ -20,8 +20,8 @@ import net.thechance.mena.designsystem.presentation.component.button.radioButton
 import net.thechance.mena.designsystem.presentation.component.text.MenaText
 import net.thechance.mena.designsystem.presentation.theme.theme.MenaTheme
 import net.thechance.mena.designsystem.presentation.theme.theme.Theme
-import net.thechance.mena.trends.presentation.screen.interestpick.CategoryScreenUiState.CategoryUiState
 import net.thechance.mena.trends.presentation.shared.component.modifier.noRippleClickable
+import net.thechance.mena.trends.presentation.shared.model.CategoryUiState
 import net.thechance.mena.trends.presentation.shared.util.Selectable
 import org.jetbrains.compose.ui.tooling.preview.Preview
