+++ trends_presentation/src/commonMain/kotlin/net/thechance/mena/trends/presentation/shared/model/CategoryUiState.kt
@@ -1,0 +1,7 @@
+package net.thechance.mena.trends.presentation.shared.model
+
+data class CategoryUiState(
+    val id: Int? = null,
+    val name: String = "",
+    val emoji: String = "",
+)
